# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "triibo",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/auth/change-password": {
      "put": {
        "tags": [
          "auth"
        ],
        "summary": "ChangePassword",
        "description": "ChangePassword",
        "operationId": "changepassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@gmail.com"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "example321"
                  },
                  "password": {
                    "type": "string",
                    "example": "example"
                  }
                }
              },
              "example": {
                "email": "example@gmail.com",
                "newPassword": "example321",
                "password": "example123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/user/email": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Email",
        "description": "Update Email",
        "operationId": "updateEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newEmail": {
                    "type": "string",
                    "example": "example@gmail.com"
                  }
                }
              },
              "example": {
                "newEmail": "example@gmail.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/user/username": {
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Username",
        "description": "Update Username",
        "operationId": "updateUsername",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "exampleUsername"
                  }
                }
              },
              "example": {
                "newUsername": "exampleUsername"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}